Vyatta-HOTFIX
Provides support for installing and listing installed packages
to a selected image.

Adds two new scripts, vyatta-hotfix and vyatta-squashfs-chroot.

Vyatta-hotfix is intended to be the hotfix input validator and chroot
executer. Vyatta-squashfs-chroot is what actually setups the overlayfs
and runs dpkg.

Dpkg is used instead of apt-get as dpkg provides an easy way to install
new packages into a partially configured chroot.

VRVDR-48759