ONIE install for UEFI: check for ESP on drive, not partition
For ONIE install, the check for the ESP fails, as fdisk is being run
for the partition instead of the drive. Fix this by passing the drive
as well as the partition to install grub. This additonal argument is
not passed on for installing grub for BIOS, but is for EFI so that it
can be used in the check for the ESP. These changes also fix a problem
with this check due to a missing whitespace for the cut.

VRVDR-49654