ONIE install for UEFI: fix chainloader for grub master
For grub2 on UEFI with GPT partitions, the chainloader command does
not understand the +1 sector map syntax which is used for BIOS. Fix
this by instead specifying the path to the relevant .efi binary. The
approach taken is due to the possibility that the ONIE-BOOT partition
might not be on the install drive used for vRouter. It was considered
more efficient to search for these partitions and use core.efi rather
than do a file search in all EFI System partitions for the relevant
grubx64.efi (this just being a copy of core.efi). Another alternative
would be to specify the configfile used instead.

VRVDR-49654