Don't install machine-id generated by the live-cd
(the live-cd does not contain a machine-id, but generates one on boot)

Some "cloning" tools like to use "clones" of freshly installed
images as if they are different from each other, so would prefer that
there is no machine-id on first boot of the image installed by the live-cd.

As long as any state that's keyed on machine-id is also not copied
into the install, this is probably is fine.

Closes: VRVDR-51179