Modelled copy command should not require prior config of ssh-known-hosts
The copy command when using an scp target requires ssh-known-hosts to
have been configured beforehand. This does not scale for deployments
with large numbers of devices and servers. The solution is to
automatically configure ssh-known-hosts when copy is used, after user
prompt acceptance. To comply with VRVDR-46681, a hash is used instead
of hostname or IP address. This presents an operational difficulty if
the public key of the server is changed, as the user cannot discern
which config to remove so as to allow the copy to subsequently succeed.
So enhance the handling also to automatically update the config for
ssh-known-hosts, again after user acceptance to a prompt.

VRVDR-50920

(cherry picked from commit 7da4666d07845e7c3a772651fc29748f51071144)