Allow concurrent removal of hosts from ssh_known_hosts
In the unlikely event that the public key of the remote server is
changed, the modelled copy command prompts the user to update this as
per VRVDR-50920. It gets the hash for the hostname/IP that is to be
deleted from the config by removing the entry in ssh_known_hosts, then
comparing this with ssh_known_hosts.old. This is valid as the current
ssh-known-hosts config does not use ssh-keygen -R or -H which result
in the .old file being created. Even if it did, the chance of this is
minimal, as it would have to happen in between the script calling
ssh-keygen -R and then getting the diff. However, a safe approach for
this was developed as part of the config changes for VRVDR-46681, so
also adopt this here by using refactored code made available in the
new SSHClient.pm module.

VRVDR-46681