Provide an option to skip host validation for the modelled copy command
The modelled copy cmd automatically updates the ssh-known-hosts config
since the fix for VRVDR-50920. In the same way as the ssh-known-hosts
fetch-from-server config, this relies on ssh-keyscan. Unlike the ssh
client itself, this does not support bind to address and so fails to
obtain the host key on setups relying on source interface capability,
due to e.g. the connected route not being present across the network or
presence of firewall. In the absence of this capability for ssh-keyscan,
provide an option to skip host validation for copying files.

VRVDR-51957