Modelled copy command in a VRF should not need prior config of ssh-known-hosts
The modelled copy command automatically configures ssh-known-hosts
since the fix for VRVDR-50920. There was no requirement for supporting
this when run in a VRF. However, an approach for this was developed as
part of the config changes for VRVDR-46681, so also adopt this here
by using refactored code made available in the new SSHClient.pm module.

VRVDR-46681