Failed to set new admin user account password during live cd image install
The installer replaces the password for the admin user account with
that provided during the live cd image install, but expects the
configuration file to have an encrypted password for the default admin
user account that needs to be replaced. However, the password in the
configuration file may also be given in plaintext, so also allow this
to be replaced by the password provided during the install.

VRVDR-49354