Image upgrade errors if config not saved
When a system image is added and config is not saved, the default
config is retrieved from the installer rootfs in /tmp so as to ensure
the config is from the new image. However, a further commit means that
the path for the default config uses the mount point for the squashfs
when installing with an ISO image file, i.e. the path is prefixed
by /mnt/cdsquash. As a result, the combined path for the default config
is incorrect, with the return due to errors meaning that there is no
prompt for new admin account credentials.

The solution is to remove the handling for retrieving the default
config from the installer rootfs, given that this file is in the
squashfs for the new ISO image. Live CD install is not affected by
this change.

VRVDR-50845