scripts: do not overlay mount on current image
vyatta-squashfs-chroot command is used for both listing
and installing packages in one of installed images. To do
this, this scripts creates an overlay mount to create the
rootfs the requested image, chroot into this new overlay and
runs dpkg/apt commands to list or install packages on this
chroot environment.

The problem happens when currently running image is the
target image. This causes the current overlay to be mounted
twice with same work and upper directory and hits an
 undefined behavior of overlayfs and kernel emits the
following log messages:

  kernel: overlayfs: upperdir is in-use as upperdir/workdir of another mount, accessing files from both mounts will
result in undefined behavior.
  kernel: overlayfs: workdir is in-use as upperdir/workdir of another mount, accessing files from both mounts will
result in undefined behavior.

The observed effect is if a file is deleted from this
overlay, the same file can't be recreated again (until
a reboot). This affects log files after logrotate. There
may be other effects that are not observed yet.

This commit fixes this by not doing a overlay mount when
the command is targeted to the currently running image.
vyatta-squashfs-chroot doesn't support installing on
currently running image, the vyatta-hotfix clones an image
and does package installation on the cloned image, but for
listing packages, we avoid remounting and run the
dpkg --list command the current root.

VRVDR-52179

(cherry picked from commit a01df3a848ffe6a74855581a662c8d2e2c6c6229)