vyatta-postinstall-configure: use rename
Moving a file across filesystem is less reliable than a rename from an
existing file on the same filesystem. For practical purposes, rename is
an atomic operation.

This commit as part of VRVDR 53096 is an improvement on the changes for
this file made under VRVDR-51837

(cherry picked from commit cf525a724d14d85817f76029200ff86b550470df)