Revert to explicit use of vyatta-curl-wrapper
So as to pass credentials to curl using environment variables as
provided by vyatta-curl-wrapper >= 0.6, re-introduce a dependency on
this package.

This is made possible due to the hard dependency on vyatta-curl-wrapper
no longer needing to be avoided, as BVNOS-bare builds were discontinued.

Note that vyatta-curl-wrapper was previously being used, given that
with update-alternatives it installs itself as /usr/bin/curl. But as
the use of environment variables for passing credentials is not
standard curl, revert to explicit use of vyatta-curl-wrapper.

VRVDR-51836