vyatta_upgrade_grub: respect kernel cmdline
If this is cloud-init image, we need to preserve the cloud-init keyword
from the kernel command line. Detect its presence and re-generate when
building a new grub configuration.

The same is true for ip=dhcp which is used in some corner cases when
generating images to support "traditional" diskless booting scenarios.

Bugfix: VRVDR-53096