vyatta_upgrade_grub: update grub.cfg when changed
When the configuration is initially loaded some of the default nodes
will trigger regenration of grub.cfg. However, the only that changes
is possibly the date in the header. If that's the case, don't replace
the existing grub.cfg. Udating critical files should be avoid if
when possible.

Bugfix: VRVDR-53096