Improve check for ONIE boot
The installer checks quite a few times whether an ONIE boot is present,
and if so exports an env var indicating this. As the result of this
check should never change during an install, as an optimization only
carry out the check once per install and then use the env var if
present.

VRVDR-53996