defer-boot: add option to not set new image as default
Allow the user to add an image but not automatically set the new image as
the default boot. This will allow the user to more easily stage upgrades.
While this could be undone by a later set system image default-boot
command, the user would need to query the running version, apply the
change and verify the desired result.

Bugfix: VRVDR-53509