Installer should check signed binaries if secure boot if enabled
If secure boot is enabled, an image with incorrectly signed binaries
results in shim, grub or kernel not coming up after boot. Recovery via
a prior good image is not possible if the shim or grub no longer come
up. A solution is for the installer to check the signed binaries of the
downloaded image against the existing signature list. This fix only
takes effect if it is in the current rather than downloaded image to be
installed, as the check is carried out before switching to the
installer of the downloaded image.

VRVDR-55074