update_grub: work around for perl module path
commit "37f8fa8  update_grub: use modules from new image" fixes the
module path so that modules could be used from the new image provided
the caller of the vyatta_update_grub.pl sets the vyatta_prefix
correctly. In older images the vyatta_prefix is not set correctly. This
fix works around the problem by ignoring the error in calls to flush_grub_cfg
subroutine.

This workaround does not completely fix all the problems of mixing
up of Perl and other libraries. Particularly, the Vyatta:Configd module
from a recent upgrade image cannot be loaded if the current image is
using an older version of libvyatta-config.so.

VRVDR-56336